refactor(stock): extract the stock engine into services
