+++ src/discordbot/services/stock/prompts.py
